refactor: extract the shared readonly-key and keyed-label concerns
