+++ src/Tables/Filters/Filter.php
@@ -10,6 +10,8 @@ use Lattice\Lattice\Attributes\SerializationHook;
 use Lattice\Lattice\Core\Components\Concerns\SerializesWireNode;
 use Lattice\Lattice\Core\Concerns\FiltersRenderableComponents;
 use Lattice\Lattice\Core\Concerns\GatesRendering;
+use Lattice\Lattice\Core\Concerns\HasKeyedLabel;
+use Lattice\Lattice\Core\Concerns\HasReadonlyKey;
 use Lattice\Lattice\Core\Contracts\Renderable;
 use Lattice\Lattice\Forms\Components\Field;
 use Lattice\Lattice\Forms\FormData;
@@ -24,32 +26,20 @@ abstract class Filter implements JsonSerializable, Renderable
 {
     use FiltersRenderableComponents;
     use GatesRendering;
+    use HasKeyedLabel;
+    use HasReadonlyKey;
     use SerializesWireNode;
-
-    public string $label;
 
     protected ?string $attribute = null;
 
     public function __construct(protected readonly string $key)
     {
-        $this->label = str($key)->headline()->toString();
+        $this->initializeLabel();
     }
 
     public static function make(string $key): static
     {
         return new static($key);
-    }
-
-    public function key(): string
-    {
-        return $this->key;
-    }
-
-    public function label(string $label): static
-    {
-        $this->label = $label;
-
-        return $this;
     }
 
     /**
@@ -68,11 +58,6 @@ abstract class Filter implements JsonSerializable, Renderable
     public function schema(): array
     {
         return [];
-    }
-
-    protected function wireKey(): ?string
-    {
-        return $this->key;
     }
 
     /**
