+++ tests/Unit/Tables/Columns/ColumnPropsTest.php
@@ -39,15 +39,18 @@ it('maps each date style method to its dateStyle/timeStyle shape', function (): 
 });
 
 it('reflects only the common concerns for columns that expose no public properties', function (): void {
-    expect(wire(BooleanColumn::make('active'))['props'])->toBe([
-        'label' => 'Active',
-        'width' => 'md',
-        'align' => 'start',
-        'sortable' => false,
-        'toggleable' => false,
-        'hiddenByDefault' => false,
-        'filter' => null,
-    ]);
+    $props = wire(BooleanColumn::make('active'))['props'];
+
+    expect($props)->toHaveCount(7)
+        ->and($props)->toMatchArray([
+            'label' => 'Active',
+            'width' => 'md',
+            'align' => 'start',
+            'sortable' => false,
+            'toggleable' => false,
+            'hiddenByDefault' => false,
+            'filter' => null,
+        ]);
 });
 
 it('keeps protected filter and sort state off the wire props', function (): void {
